Rename to stackhpc-cloud-tests
