+++ README.md
@@ -1,6 +1,6 @@
-# StackHPC OpenStack Tests
+# StackHPC Cloud Tests
 
-Automated testing for StackHPC OpenStack.
+Automated testing for StackHPC OpenStack Clouds.
 
 Provides test coverage of various aspects of OpenStack and related services, including:
 
@@ -19,7 +19,7 @@ Create a virtual environment.
 python3 -m venv venv
 ```
 
-Install stackhpc-openstack-tests and its dependencies.
+Install stackhpc-cloud-tests and its dependencies.
 
 ```sh
 venv/bin/pip install <path/to/repo> -r <path/to/repo>/requirements.txt
@@ -30,11 +30,11 @@ venv/bin/pip install <path/to/repo> -r <path/to/repo>/requirements.txt
 Run all tests provided.
 
 ```sh
-py.test --pyargs stackhpc_openstack_tests
+py.test --pyargs stackhpc_cloud_tests
 ```
 
 Or run tests from a specific submodule.
 
 ```sh
-py.test --pyargs stackhpc_openstack_tests.test_prometheus
+py.test --pyargs stackhpc_cloud_tests.test_prometheus
 ```
