+++ setup.cfg
@@ -1,11 +1,11 @@
 [metadata]
-name = stackhpc-openstack-tests
-version = 0.1.0
-summary = Automated testing for StackHPC OpenStack
+name = stackhpc-cloud-tests
+version = 0.2.0
+summary = Automated testing for StackHPC OpenStack Clouds
 description-file = README.md
 author = Mark Goddard
 author-email = mark@stackhpc.com
-url = https://github.com/stackhpc/stackhpc-openstack-tests
+url = https://github.com/stackhpc/stackhpc-cloud-tests
 python-requires = >=3.6
 classifier =
     Environment :: OpenStack
@@ -26,7 +26,7 @@ classifier =
 
 [files]
 packages =
-    stackhpc_openstack_tests
+    stackhpc_cloud_tests
 
 [options]
 install_requires =
