+++ stackhpc_cloud_tests/utils.py
